+++ src/scenes/components/menuItemRow.c
@@ -112,7 +112,7 @@ bool IsMenuItemRowHovered(const MenuItemRow_t* item) {
     int my = app.mouse.y;
 
     // Calculate text dimensions for accurate bounds
-    int text_w, text_h;
+    float text_w, text_h;
     a_CalcTextDimensions((char*)item->label, FONT_ENTER_COMMAND, &text_w, &text_h);
 
     // Calculate bounds: x centered, y aligned with text rendering
@@ -161,7 +161,7 @@ void RenderMenuItemRow(const MenuItemRow_t* item) {
     mutable_item->is_hovered = IsMenuItemRowHovered(item);
 
     // Calculate text dimensions for accurate bounds
-    int text_w, text_h;
+    float text_w, text_h;
     a_CalcTextDimensions((char*)item->label, FONT_ENTER_COMMAND, &text_w, &text_h);
 
     // Draw hover background (semi-transparent white rectangle)
